Improve char width computation (#593)
Char width computation didn't work as expected when the first character
of the NSTextView was an emoji. The `maximumAdvancement` returned the
length of the emoji, placing the margins at unexpected position.

This commit improves the computation by using a fixed character (`x` has
been chosen, arbitrarily). The margin will thus be placed at 50
non-emoji characters.